ui/rsl: improved trip tooltip
